requirements.txt
streamlit
llama-index
llama-index-embeddings-huggingface
llama-index-llms-huggingface
llama-index-vector-stores-faiss
sentence-transformers
transformers
torch
pypdf
faiss-cpu
